feat: Pipeline name resolution
We remove internal name key support for the blueprint files, and just
use their filename-stems as keys. This will greatly simplify the
iteration processes to both find, lookup and update blueprints.

And since most operating systems do not love identical file names, we
can offload the proof of uniqueness to the OS :^)